Moved constants to constants.js
