+++ www/src/components/shell/constants.js
@@ -1,1 +1,35 @@
 export const CLOUDS = ['AWS', 'GCP']
+
+export const SECTION_GIT = 'git'
+export const SECTION_CLOUD = 'cloud'
+export const SECTION_WORKSPACE = 'workspace'
+export const SECTION_FINISH = 'finish'
+export const SECTION_INSTALL_CLI = 'cli'
+
+export const SECTIONS = {
+  [SECTION_GIT]: {
+    next: SECTION_CLOUD,
+    previous: null,
+    stepIndex: 0,
+  },
+  [SECTION_CLOUD]: {
+    next: SECTION_WORKSPACE,
+    previous: SECTION_GIT,
+    stepIndex: 1,
+  },
+  [SECTION_WORKSPACE]: {
+    next: SECTION_FINISH,
+    previous: SECTION_CLOUD,
+    stepIndex: 2,
+  },
+  [SECTION_FINISH]: {
+    next: null,
+    previous: SECTION_WORKSPACE,
+    stepIndex: 3,
+  },
+  [SECTION_INSTALL_CLI]: {
+    next: null,
+    previous: null,
+    stepIndex: 1,
+  },
+}
